requirements.txt
# Core Dependencies
groq
google-generativeai>=0.4.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# UI
pywebview>=4.0.0
pystray>=0.19.0
Pillow>=10.0.0

# Speech
SpeechRecognition>=3.10.0
PyAudio>=0.2.13
pyttsx3>=2.90

# Memory
mem0ai>=0.0.1
qdrant-client>=1.7.0

# Vision
mss>=10.0.0
pytesseract>=0.3.0
easyocr>=1.7.0

# Control
pyautogui>=0.9.54
keyboard>=0.13.5
playwright>=1.40.0
pyperclip>=1.8.2
